size_t problem solved

+++ src/Access.cpp
@@ -77,7 +77,7 @@ bool Access::OpenDevice(string dev_name) {
 
 }
 
-bool Access:: Write_on_device(char* note,int byte_offset,int size){
+bool Access:: Write_on_device(char* note,int byte_offset,size_t size){
 
     //////Setting the pointer to where to write the file
     error_code = SetFilePointer(Disk_101, (byte_offset), 0, FILE_BEGIN);
@@ -105,7 +105,7 @@ bool Access:: Write_on_device(char* note,int byte_offset,int size){
 }
 
 
-bool Access:: Read_from_device(char* note,int byte_offset,int size){
+bool Access:: Read_from_device(char* note,int byte_offset,size_t size){
     // Read the data back from the disk device
     //////Setting the pointer to where to start reading the file
     error_code = SetFilePointer(Disk_101, byte_offset, 0, FILE_BEGIN);
